- rpc listspentutxo

+++ transaction/txout.go
@@ -1,11 +1,11 @@
 package transaction
 
 type TxOut struct {
-	Value    int
+	Value    float64
 	PkScript []byte
 }
 
-func (self TxOut) NewTxOut(value int, pkScript []byte) *TxOut {
+func (self TxOut) NewTxOut(value float64, pkScript []byte) *TxOut {
 	self = TxOut{
 		Value:    value,
 		PkScript: pkScript,
